analysis: split neutral model and per-format adapters into separate files
model.go previously held the neutral model, the pprof adapter, the OTLP adapter
and the format dispatch all at once. Separate them by concern so each format
stands alone and is easy to read:

- model.go  - format-agnostic core: ProfileSet, canonical labels + canonKey,
              and LoadProfileSet dispatch.
- pprof.go  - FromPprof (+ label/stack folding helpers).
- otlp.go   - FromOTLP, with the dictionary walking wrapped in a small
              bounds-checked otlpDict resolver (str/funcName/mappingBase/
              foldStack/sampleLabels) instead of a pile of inline closures, so
              FromOTLP itself is a short, readable loop.

Tests split to match: pprof_test.go, otlp_test.go, and model_test.go (canonKey
+ loader dispatch). No behavior change; all tests pass, gofmt/ruff clean.
